+++ R/get_peaks.R
@@ -5,9 +5,11 @@
 #' 
 #' Peaks are located by finding zero-crossings in the smoothed first derivative
 #' of the specified chromatographic traces (function \code{\link{find_peaks}}).
-#' At the given positions, an exponential-gaussian hybrid (or regular gaussian)
-#' function is fit to the signal using \code{\link{fit_peaks}} according to the
-#' value of \code{fit}. Finally, the area is calculated using trapezoidal 
+#' At the given positions, a peak model is fit to the signal using 
+#' \code{\link{fit_peaks}} according to the value of the \code{fit} argument.
+#' Available models include the bidirectional exponentially modified gaussian
+#' (\code{bemg}),exponential-gaussian hybrid (\code{egh} or regular 
+#' \code{gaussian}.  Finally, the area is calculated using a trapezoidal 
 #' approximation.
 #'
 #' Additional arguments can be provided to \code{\link{find_peaks}} to fine-tune
@@ -16,7 +18,8 @@
 #' aggressive smoothing may cause small peaks to be overlooked. 
 #' 
 #' The standard deviation (\code{sd}), full-width at half maximum (\code{FWHM}),
-#' tau \code{tau}, and \code{area} are returned in units determined by 
+#' \eqn{\tau} (\code{tau}), \eqn{\tau_{R}} (\code{tau_right}), 
+#' \eqn{\tau_{L}} (\code{tau_left}), and \code{area} are returned in units determined by 
 #' \code{time_unit}. By default, the units are in minutes. To compare directly
 #' with 'ChemStation' integration results, the time units should be changed to
 #' seconds.
@@ -27,10 +30,13 @@
 #' (timepoints × wavelengths).
 #' @param lambdas A character or numeric vector specifying the wavelengths to 
 #' find peaks at. For one-dimensional chromatograms, this argument can be ignored.
-#' @param fit What type of fit to use. Current options are exponential-gaussian
-#' hybrid (\code{egh}), gaussian or raw. The \code{raw} setting performs
-#' trapezoidal integration directly on the raw data without fitting a peak shape.
-#' @param sd_max Maximum width (standard deviation) for peaks. Defaults to 50.
+#' @param fit What type of fit to use. Current options are bidirectional 
+#' exponentially modified gaussian (\code{bemg}), exponential-gaussian
+#' hybrid (\code{egh}), \code{gaussian}, or \code{raw}. The \code{raw} setting 
+#' performs trapezoidal integration directly on the raw data without fitting a 
+#' model to the data. Defaults to \code{bemg}.
+#' @param sd_max Maximum width (standard deviation) for peaks. Defaults to 
+#' \code{50}.
 #' @param sd.max The \code{sd.max} argument is deprecated. Please use 
 #' \code{sd_max} instead.
 #' @param max_iter Maximum number of iterations for non-linear least squares
@@ -66,11 +72,16 @@
 #' * `rt`: The retention time of the peak maximum.
 #' * \code{start}: The retention time where the peak is estimated to begin.
 #' * \code{end}: The retention time where the peak is estimated to end.
-#' * \code{sd}: The standard deviation of the fitted peak shape.
-#' * \code{tau} The value of parameter \eqn{\tau}. This parameter determines 
-#' peak asymmetry for peaks fit with an exponential-gaussian hybrid function.
-#' (This column will only appear if \code{fit = egh}.
-#' * \code{FWHM}: The full-width at half maximum.
+#' * \code{sd}: The standard deviation (\eqn{\sigma}) of the fitted peak shape.
+#' * \code{tau} Empirical skewness parameter (in units of time) determining peak
+#' asymmetry when \code{fit = "egh"}.
+#' * \code{tau_right} Exponential rate constant controlling right-sided tailing
+#' when \code{fit = "bemg"}. Note that unlike \code{tau} in the EGH model,
+#' this parameter has units of 1/time.
+#' * \code{tau_left} Exponential rate constant controlling left-sided tailing
+#' when \code{fit = "bemg"}. Note that unlike \code{tau} in the EGH model,
+#' this parameter has units of 1/time.
+#' * \code{FWHM}: The full-width at half maximum calculated as \eqn{2.355 \sigma}.
 #' * \code{height}: The height of the peak.
 #' * \code{area}: The area of the peak as determined by trapezoidal approximation.
 #' * \code{r.squared} The coefficient of determination (\eqn{R^2}) of the fitted
@@ -90,9 +101,15 @@
 #' functions as a simple model of asymmetric chromatographic peaks. \emph{Journal of
 #' Chromatography A} \strong{915}:1-13. \doi{10.1016/S0021-9673(01)00594-5}.
 #'
-#' * Naish, P. J. & Hartwell, S. 1988. Exponentially Modified Gaussian functions - A
-#' good model for chromatographic peaks in isocratic HPLC? \emph{Chromatographia},
+#' * Naish, P. J. & Hartwell, S. 1988. Exponentially Modified Gaussian functions 
+#' — A good model for chromatographic peaks in isocratic HPLC? \emph{Chromatographia},
 #' \strong{26}: 285-296. \doi{10.1007/BF02268168}.
+#' 
+#' * Arase, Shuntaro, Kanta Horie, Takashi Kato, et al. 2016. Intelligent Peak 
+#' Deconvolution through In-Depth Study of the Data Matrix from Liquid 
+#' Chromatography Coupled with a Photo-Diode Array Detector Applied to 
+#' Pharmaceutical Analysis. *Journal of Chromatography. A*, **1469**: 35–47. 
+#' \doi{doi.org/10.1016/j.chroma.2016.09.037}.
 #'
 #' * O'Haver, Tom. Pragmatic Introduction to Signal Processing:
 #' Applications in scientific measurement.
@@ -109,7 +126,8 @@
 #' @export get_peaks
 #' @md
 
-get_peaks <- function(chrom_list, lambdas, fit = c("egh", "gaussian", "raw"),
+get_peaks <- function(chrom_list, lambdas, 
+                      fit = c("bemg", "egh", "gaussian", "raw"),
                       sd_max = 50, max_iter = 100,
                       time_unit = c("min", "s", "ms"),
                       estimate_purity = FALSE,  noise_threshold = .001,
@@ -120,7 +138,7 @@ get_peaks <- function(chrom_list, lambdas, fit = c("egh", "gaussian", "raw"),
   time_unit <- resolve_deprecated(time.units, time_unit)
   time_unit <- match.arg(time_unit, c("min", "s", "ms"))
   tfac <- switch(time_unit, "min" = 1, "s" = 60, "ms" = 60*1000)
-  fit <- match.arg(fit, c("egh", "gaussian", "raw"))
+  fit <- match.arg(fit, c("bemg", "egh", "gaussian", "raw"))
   chrom_list_string <- deparse(substitute(chrom_list))
   if (class(chrom_list)[1] == "matrix")
     chrom_list <- list(chrom_list)
@@ -194,6 +212,9 @@ convert_indices_to_times <- function(x, chrom_list, idx, tfac){
   x[, c('sd', 'FWHM', 'area')] <- x[, c('sd', 'FWHM', 'area')] * tdiff * tfac
   if (!is.null(x$tau)){
     x[, c('tau')] <- x[, c('tau')] * tdiff * tfac
+  } 
+  if (!is.null(x$tau_right)){
+    x[, c('tau_right', 'tau_left')] <- x[, c('tau_right', 'tau_left')] / (tdiff * tfac)
   } 
   x
 }