+++ R/plot.peak_list.R
@@ -11,18 +11,20 @@
 #' matrices)
 #' @param idx Index or name of chromatogram to be plotted.
 #' @param index This argument is deprecated. Please use \code{idx} instead.
-#' @param lambda Wavelength for plotting.
-#' @param points Logical. If TRUE, plot peak maxima. Defaults to FALSE.
-#' @param ticks Logical. If TRUE, mark beginning and end of each peak. Defaults
-#' to FALSE.
-#' @param a Alpha parameter controlling the transparency of fitted shapes.
+#' @param lambda Wavelength to use for plotting.
+#' @param points Logical. If \code{TRUE}, plot peak maxima. Defaults to 
+#' \code{FALSE}.
+#' @param ticks Logical. If \code{TRUE}, mark beginning and end of each peak. 
+#' Defaults to \code{FALSE}.
+#' @param alpha Parameter controlling the transparency of fitted shapes.
+#' Defaults to \code{0.5}.
 #' @param color The color of the fitted shapes.
-#' @param cex.points Size of points. Defaults to 0.5
-#' @param numbers Whether to number peaks. Defaults to FALSE.
-#' @param cex.font Font size if peaks are numbered. Defaults to 0.5.
-#' @param y.offset Y offset for peak numbers. Defaults to 25.
+#' @param cex.points Size of points. Defaults to \code{0.5}.
+#' @param numbers Whether to number peaks. Defaults to \code{FALSE}.
+#' @param cex.font Font size if peaks are numbered. Defaults to \code{0.5}.
+#' @param y.offset Y offset for peak numbers. Defaults to \code{25}.
 #' @param plot_purity Whether to add visualization of peak purity.
-#' @param res time resolution for peak fitting
+#' @param res Time resolution for peak fitting
 #' @return No return value, called for side effects.
 #' @section Side effects:
 #' Plots a chromatographic trace from the specified chromatogram (\code{chr})
@@ -39,7 +41,7 @@
 #' @export
 
 plot.peak_list <- function(x, ..., chrom_list, idx = 1, lambda = NULL,
-                           points = FALSE, ticks = FALSE, a = 0.5, color = NULL,
+                           points = FALSE, ticks = FALSE, alpha = 0.5, color = NULL,
                            cex.points = 0.5, numbers = FALSE, cex.font = 0.5, 
                            y.offset = 25, plot_purity = FALSE, res, index = NULL){
   if (!is.null(index)){
@@ -64,11 +66,7 @@ plot.peak_list <- function(x, ..., chrom_list, idx = 1, lambda = NULL,
   new.ts <- get_times(x = chrom_list, idx = idx)
   y <- chrom_list[[idx]][,lambda]
   pks <- data.frame(x[[idx]][[lambda]])
-  if ("r.squared" %in% colnames(pks)){
-    fit <- ifelse("tau" %in% colnames(pks), "egh", "gaussian")
-  } else{
-    fit <- "raw"
-  }
+  fit <- attr(x,"fit")
   plot(new.ts, y, type = 'l', xlab = '', ylab = '', xaxt = 'n', yaxt = 'n', ...)
   if (points){
     points(pks$rt, pks$height, pch = 20, cex=cex.points, col = 'red')
@@ -76,10 +74,10 @@ plot.peak_list <- function(x, ..., chrom_list, idx = 1, lambda = NULL,
   if (ticks){
     arrows(pks$start, y[which(new.ts %in% pks$start)] - 5,
            pks$start, y[which(new.ts %in% pks$start)] + 5,
-           col="blue", length = 0)
+           col = "blue", length = 0)
     arrows(pks$end, y[which(new.ts %in% pks$end)] - 5,
            pks$end,y[which(new.ts %in% pks$end)] + 5,
-           col="blue", length = 0)
+           col = "blue", length = 0)
   }
   if (numbers){
     text(pks$rt, y[pks$rt] + y.offset, labels = seq_len(nrow(pks)), 
@@ -89,26 +87,28 @@ plot.peak_list <- function(x, ..., chrom_list, idx = 1, lambda = NULL,
     res <- get_time_resolution(chrom_list)
   for (i in seq_len(nrow(pks))){
     try({
-      peak.loc<-seq.int((pks$start[i]),(pks$end[i]), by = res)
+      peak.loc <- seq.int((pks$start[i]),(pks$end[i]), by = res)
       if (fit == "gaussian"){
         yvals <- gaussian(peak.loc, center=pks$rt[i],
                           width=pks$sd[i]*tfac, height = pks$height[i])
-        if (is.null(color))
-          color <- "red"
+        if (is.null(color)) color <- "red"
       }
       else if (fit == "egh"){
         yvals <- egh(x = peak.loc, center = pks$rt[i],
                      width=pks$sd[i]*tfac, height = pks$height[i],
                      tau = pks$tau[i]*tfac)
-        if (is.null(color))
-          color <- "purple"
+        if (is.null(color)) color <- "purple"
+      } else if (fit == "bemg"){
+        yvals <- bemg(x = peak.loc, center = pks$rt[i], width = pks$sd[i]*tfac,
+                      height = pks$height[i], a = pks$tau_right[i]*tfac, 
+                      b = pks$tau_left[i]*tfac)
+        if (is.null(color)) color <- "darkgreen"
       }
       else if (fit == "raw"){
         yvals <- chrom_list[[idx]][as.character(peak.loc), lambda]
-        if (is.null(color))
-          color <- "hotpink"
+        if (is.null(color)) color <- "hotpink"
       }
-      draw_trapezoid(peak.loc, yvals, color, a)
+      draw_trapezoid(peak.loc, yvals, color, alpha = alpha)
     }, silent = TRUE)
   }
   if (plot_purity){
@@ -125,7 +125,7 @@ plot.peak_list <- function(x, ..., chrom_list, idx = 1, lambda = NULL,
         p <- get_purity_values(chrom_list[[idx]], pos)
         lim = -20
         draw_trapezoid(new.ts[pk_indices], scales::rescale(p, c(0, lim)), 
-                       color = "black", a = 0.6)
+                       color = "black", alpha = 0.6)
         abline(h = lim, lty = 3, col = "darkgray")
       })
     }, silent = TRUE)
@@ -133,9 +133,9 @@ plot.peak_list <- function(x, ..., chrom_list, idx = 1, lambda = NULL,
 }
 
 #' @noRd
-draw_trapezoid <- function(peak.loc, yvals, color, a){
+draw_trapezoid <- function(peak.loc, yvals, color, alpha){
   sapply(seq_len(length(peak.loc) - 1), function(i){
     polygon(peak.loc[c(i, i, (i + 1), (i + 1))], c(0, yvals[i:(i + 1)], 0),
-            col = scales::alpha(color, a), lty = 3, border = NA)
+            col = scales::alpha(color, alpha), lty = 3, border = NA)
   })
 }